Added Washington State Redistricting Project
In this project, we worked on creating an optimized redistricting plan for the state of Washington using the 2020 US Census data. Our goal was to address the challenges of population distribution and racial balance across the state's districts. We defined an objective function to ensure equal representation by assigning one representative per district, while also considering racial and population factors. To achieve this, we developed constraints based on county adjacency and set covering, which helped us create a redistricting solution that aims to be fair and equitable, balancing racial demographics and ensuring proportional representation across the state.